**environ and posix_spawnp
Exploration on how to spawn other processes. Exposure to char **environ variable that holds the visible environment variables and how to used it in the program. StartAuxProcesses() is written to test spawning a simple process. The output of the spawned command is in the console. I am sure redirects will work as command line arguments